sanders: Add DT2W Support
Sanders (Moto G5S Plus) uses FocalTech Touch driver (ft5x06).
This DT2W is picked from the work I did for Redmi2, in which I fixed
many bugs including
 * irq imbalance
 * hardware keys disable/enable
 * display panel changes to handle screen on/off states

It is expected that proximity wake is supported by the ROM, so it would
be complemented with dt2w. Otherwise, we need to implement ondemand
proximity sensor which I have done for Redmi2, but no thanks, do not
want to do for sanders as well.

We can use wake-gestures, which includes sweep2wake etc, too but i
prefer this old style of just dt2w implemenation which i m comfortable
with.